skysky

+++ sky_pattern/sky_template_taskfarmer/compute_sky_templates_v2.py
@@ -80,20 +80,30 @@ ccd_columns = ['image_hdu', 'expnum', 'ccdname', 'ccdskycounts']
 ccd = Table(fitsio.read(surveyccd_path, columns=ccd_columns))
 print(len(ccd))
 
-# skyrun = Table.read('/global/cscratch1/sd/rongpu/temp/skyrunsgoodcountexpnumv48dr8.fits')
-skyrun = Table.read('/global/cscratch1/sd/rongpu/temp/skyrunsgoodcountexpnumv48dr8_less.fits')
+skyrun = Table.read('/global/cscratch1/sd/rongpu/temp/skyrunsgoodcountexpnumv48dr8.fits')
+# skyrun = Table.read('/global/cscratch1/sd/rongpu/temp/skyrunsgoodcountexpnumv48dr8_less.fits')
 print(len(skyrun))
 
 mask = skyrun['ok']==True
 skyrun = skyrun[mask]
 print(len(skyrun))
 
+# Exclude templates already created
+fn_list = glob.glob(os.path.join(output_dir, '*.fits.fz'))
+run_list_done = [int(fn[len(os.path.join(output_dir, 'sky_templates_'))+1:-8]) for fn in fn_list]
+mask = ~np.in1d(skyrun['run'], run_list_done)
+skyrun = skyrun[mask]
+print(len(skyrun), len(run_list_done))
+
+# ########################## Exclude z band ##########################
 # band = 'z'
-# mask = skyrun['filter']==band
+# mask = skyrun['filter']!=band
 # skyrun = skyrun[mask]
 # print(len(skyrun))
+# ####################################################################
 
 run_list = np.unique(skyrun['run'])
+print(len(run_list))
 
 # shuffle
 np.random.seed(123)
@@ -137,7 +147,7 @@ def compute_smooth_sky(run, plot_q=False, diagnostic_touch=True):
         print(ccdnum)
 
         # ####################
-        # start = time.clock()
+        # start = time.time()
         # ####################
 
         img_list = []
@@ -187,7 +197,8 @@ def compute_smooth_sky(run, plot_q=False, diagnostic_touch=True):
             # ccd_index = np.where((ccd['expnum']==skyrun['expnum'][skyrun_index]) & (ccd['image_hdu']==hdu_index))[0][0]
 
             # Get the median ccdskycounts of the exposure
-            mask = (ccd['expnum']==skyrun['expnum'][skyrun_index]) & (ccd['ccdname']!='S7') & (ccd['ccdname']!='S7 ')
+            # mask = (ccd['expnum']==skyrun['expnum'][skyrun_index]) & (ccd['ccdname']!='S7') & (ccd['ccdname']!='S7 ') # too slow!
+            mask = ccd['expnum']==skyrun['expnum'][skyrun_index]
             ccdskycounts_median = np.median(ccd['ccdskycounts'][mask])
             # print('ccdskycounts_median = {:.4f}'.format(ccdskycounts_median))
             
@@ -299,13 +310,14 @@ def compute_smooth_sky(run, plot_q=False, diagnostic_touch=True):
         hdul_w.write(data=img_median_smooth, extname=ccdname, compress='rice')
 
         # ##################
-        # end = time.clock()
+        # end = time.time()
         # print('Took {:.1f} seconds'.format(end - start))
         # ##################
 
     hdul_w.close()
     
-    os.remove('/global/u2/r/rongpu/temp/sky_template_being_written/'+os.path.basename(output_path))
+    if diagnostic_touch:
+        os.remove('/global/u2/r/rongpu/temp/sky_template_being_written/'+os.path.basename(output_path))
 
 def main():
